회원가입 디자인 수정
로그인 과정 중간에 추가 액티비티 생성
