+++ app/src/main/AndroidManifest.xml
@@ -1,25 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.abbsolute.ma_livu">
-
-    <!--인터넷 허가-->
-    <uses-permission android:name="android.permission.INTERNET"/>
+    <!-- 인터넷 허가 -->
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:name=".GlobalApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:name=".GlobalApplication"
         android:theme="@style/AppTheme">
-
-        <!--GlobalApplication 객체를 공유하기 위함-->
+        <activity android:name=".Login2Activity"></activity>
+        <!-- GlobalApplication 객체를 공유하기 위함 -->
         <meta-data
             android:name="com.kakao.sdk.AppKey"
             android:value="@string/kakao_app_key" />
 
-        <activity android:name=".CommuWriteActivity"></activity>
+        <activity android:name=".CommuWriteActivity" />
         <activity android:name=".CommunityActivity" />
         <activity android:name=".HomeActivity" />
         <activity android:name=".SignupActivity" />
